requirements.txt
annotated-types==0.7.0
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.0
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
setuptools==74.1.2
typing_extensions==4.12.2
